+++ server/src/controllers/Comment/deleteComment.ts
@@ -1,4 +1,4 @@
-import { Comment } from "../db";
+import { Comment } from "../../db";
 
 export const deleteComments = async (id: string) => {
   const deletedComment = await Comment.findByPk(id);
