+++ server/src/controllers/Comment/getComment.ts
@@ -1,4 +1,4 @@
-import { Comment } from "../db";
+import { Comment } from "../../db";
 
 export const getComments = async (productId: string) => {
   try {
