+++ server/src/controllers/Product/getProducts.ts
@@ -1,0 +1,7 @@
+import { Product } from "../../db";
+
+
+// Controlador para obtener todos los productos
+export async function getProducts() {
+    return Product.findAll();
+  }
