+++ server/src/handlers/order/getOrders.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { getOrders } from "../../controllers/order/getOrders";
+import { getOrders } from "../../controllers/Order/getOrders";
 
 export async function getOrderHandler(req: Request, res: Response) {
     try {
