+++ server/src/handlers/user/updateUser.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { updateUser } from '../../controllers/user/updateUser';
+import { updateUser } from '../../controllers/User/updateUser';
 import { User } from "../../db";
 
 export async function updateUserHandler(req: Request, res: Response) {
