--- server/src/models/Favourites.ts
@@ -1,20 +1,0 @@
-import { DataTypes } from 'sequelize';
-import { Sequelize } from 'sequelize-typescript';
-
-export default function initializeFavouritesModel(sequelize: Sequelize): void {
-  sequelize.define('Favourite',
-    {
-      id: {
-        type: DataTypes.INTEGER,
-        autoIncrement: true,
-        primaryKey: true
-      },
-      productId: {
-        type: DataTypes.INTEGER,
-      }
-    },
-    {
-      timestamps: false,
-    }
-  )
-}
