--- server/src/routes/favourites.ts
@@ -1,13 +1,0 @@
-/* import express from "express";
-import { getFavouriteHandler } from "../handlers/favourite/getFavourite";
-import { createFavouriteHandler } from "../handlers/favourite/postFavourite";
-import { deleteFavouriteHandler } from "../handlers/favourite/deleteFavourite";
-
-const favoritesRouter =express.Router();
-
-favoritesRouter.get("/:id",getFavouriteHandler);
-favoritesRouter.post("/",createFavouriteHandler);
-favoritesRouter.delete("/:id",deleteFavouriteHandler);
-
-
-export default favoritesRouter; */
