--- server/src/routes/status.ts
@@ -1,13 +1,0 @@
-import express from "express";
-import { createStatusHandler, deleteStatusHandler, getStatusByIdHandler, getStatusesHandler, updateStatusHandler } from "../handlers/statusHandlers";
-
-const statusRouter = express.Router();
-
-statusRouter.post("/",createStatusHandler);
-statusRouter.get("/:id",getStatusByIdHandler);
-statusRouter.get("/",getStatusesHandler);
-statusRouter.put("/:id",updateStatusHandler);
-statusRouter.delete("/:id",deleteStatusHandler);
-
-
-export default statusRouter;
